Add files via upload
Using functions in python and passing parameters is an another method of passing different values to specific operations. In this program, the max_num function takes three numbers as input and returns the largest of the three. We can use this function to find the maximum value among any set of numeric inputs.