Add files via upload
Here, I have written a basic code for months and their funfacts with the help of python dictionaries.
A Python dictionary is a collection of key-value pairs where each key maps to a specific value. It allows for fast data retrieval using unique keys, making it ideal for tasks like mapping, storing, and organizing data. Dictionaries are created using curly braces {} and are dynamic, meaning they can be easily modified and expanded.